refactor: remove unused MeshTenantSorting and move InteractivePrompts
